requirements.txt
asgiref==3.7.2
blinker==1.6.2
boto3==1.28.53
botocore==1.31.53
Django==4.2.5
django-environ==0.11.2
djangorestframework==3.14.0
dynamorm==0.11.0
jmespath==1.0.1
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.16
Werkzeug==2.3.7
